requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-google-genai
google-ai-generativelanguage
python-dotenv
pydantic
duckduckgo-search